+++ native/include/HttpServerWrapper.h
@@ -7,7 +7,7 @@ class HttpServerWrapper : public ComObject<IHttpServer, &IID_IHttpServer>
     public:
         HttpServerWrapper(ProwingenRequestHandler handler);
         virtual ~HttpServerWrapper();
-        virtual HRESULT AddAddress(char*host, int port, bool lookup) override;
+        virtual HRESULT AddAddress(char*host, uint16_t port, bool lookup) override;
         virtual HRESULT Start() override;
     protected:
     private:
